Updated Renderable.java
There are no more errors in the file, however I expect some problems
when looping through the _properties list with all of the different
objects with various key-values types.

Also, this version is still using ENUMSET, so I will be looking into
changing this.

Finally, I changed the switch statement to ifs because it was
complaining about the values not being constant.